Crate fil_types

Structs

OnChainSealVerifyInfo
OnChainSealVerifyInfo is the structure of information that must be sent with a message to commit a sector. Most of this information is not needed in the state tree but will be verified in sm.CommitSector. See SealCommitment for data stored on the state tree for each sector.

OnChainWindowPoStVerifyInfo
Information submitted by a miner to provide a Window PoSt.

PaddedPieceSize
Size of a piece in bytes with padding

PieceInfo

PoStProof

SealVerifyInfo
Information needed to verify a seal proof.

SectorID
Sector ID which contains the sector number and the actor ID for the miner.

SectorInfo
Information about a sector necessary for PoSt verification

UnpaddedPieceSize
Size of a piece in bytes

WindowPoStVerifyInfo
Information needed to verify a Window PoSt submitted directly to a miner actor.

WinningPoStVerifyInfo
Information needed to verify a Winning PoSt attached to a block header. Note: this is not used within the state machine, but by the consensus/election mechanisms.

Enums

RegisteredProof
This ordering, defines mappings to UInt in a way which MUST never change.

SectorSize
SectorSize indicates one of a set of possible sizes in the network.

Functions

zero_piece_commitment

Type Definitions

InteractiveSealRandomness

PoStRandomness

SealRandomness

SectorNumber

StoragePower
Unit of storage power (measured in bytes)