Module fil_types::build_version

Structs

APIVersion

Represents the current version of the API.

BUILD_TYPE

CURRENT_COMMIT

RUNNING_NODE_TYPE

Version

Integer based value on version information. Highest order bits for Major, Mid order for Minor and lowest for Patch.

Enums

BuildType

Build type for the node. This shares which build type the node is from the RPC API.

NodeType

The type of node that is running.

Functions

user_version

Gets the formatted current user version.