Trait fil_types::verifier::ProofVerifier

pub trait ProofVerifier {
    fn verify_seal(vi: &SealVerifyInfo) -> Result<(), Box<dyn StdError>> { ... }

    fn verify_winning_post(
        Randomness: Randomness, 
        proofs: &[PoStProof], 
        challenge_sectors: &[SectorInfo], 
        prover: u64
    ) -> Result<(), Box<dyn StdError>> { ... }

    fn verify_window_post(
        Randomness: Randomness, 
        proofs: &[PoStProof], 
        challenge_sectors: &[SectorInfo], 
        prover: u64
    ) -> Result<(), Box<dyn StdError>> { ... }

    fn generate_winning_post_sector_challenge(
        proof: RegisteredPoStProof, 
        prover_id: u64, 
        Randomness: Randomness, 
        eligible_sector_count: u64
    ) -> Result<Vec<u64>, Box<dyn StdError>> { ... }
}

Functionality for verification of seal, winning PoSt and window PoSt proofs. Proof verification will be full validation by default.

Provided methods

fn verify_seal(vi: &SealVerifyInfo) -> Result<(), Box<dyn StdError>>

Verify seal proof for sectors. This proof verifies that a sector was sealed by the miner.

fn verify_winning_post(
Randomness: Randomness,
proofs: &[PoStProof],
challenge_sectors: &[SectorInfo],
prover: u64
) -> Result<(), Box<dyn StdError>>

Verifies winning proof of spacetime. These proofs are generated by the miners that are elected to mine a new block to verify a sector. A failed winning proof leads to a miner being slashed.

fn verify_window_post(
Randomness: Randomness,
proofs: &[PoStProof],
challenge_sectors: &[SectorInfo],
prover: u64
) -> Result<(), Box<dyn StdError>>

Verifies window proof of spacetime. These proofs are generated regularly to audit commitments made by storage miners. An invalid Window PoSt leads to a miner's block being invalidated and miss the opportunity to receive block rewards.

fn generate_winning_post_sector_challenge(
proof: RegisteredPoStProof,
prover_id: u64,
Randomness: Randomness,
eligible_sector_count: u64
) -> Result<Vec<u64>, Box<dyn StdError>>

Generates sector challenge indexes for use in winning PoSt verification.

Loading content...