Crate fil_types

Re-exports

pub use self::sector::*;

Modules

build_version

deadlines

genesis

sector

verifier

Structs

DefaultNetworkParams
Params for the network. This is now continued on into mainnet and is static across networks.

FIL_RESERVED
Amount of total Filecoin reserved in a static ID address.

PaddedPieceSize
Size of a piece in bytes with padding.

PieceInfo
Piece information for part or a whole file.

Randomness
String of random bytes usually generated from a randomness beacon or from tickets on chain.

StateInfo0
Empty state tree information. This is serialized as an array for future proofing.

StateRoot
State root information. Contains information about the version of the state tree, the root of the tree, and a link to the information about the tree.

TOTAL_FILECOIN
Total Filecoin available to the network.

UnpaddedPieceSize
Size of a piece in bytes.

ZERO_ADDRESS
Zero address used to avoid allowing it to be used for verification. This is intentionally disallowed because it is an edge case with Filecoin’s BLS signature verification.

Enums

NetworkVersion
Specifies the network version

StateTreeVersion
Specifies the version of the state tree

Constants

ALLOWABLE_CLOCK_DRIFT
Allowable clock drift in validations.

BLOCKS_PER_EPOCH
The expected number of block producers in each epoch.

BLOCK_GAS_LIMIT
Total gas limit allowed per block. This is shared across networks.

FILECOIN_PRECISION
Ratio of integer values to token value.

HAMT_BIT_WIDTH
Default bit width for the hamt in the filecoin protocol.

TICKET_RANDOMNESS_LOOKBACK
Lookback height for retrieving ticket randomness.

TOTAL_FILECOIN_BASE
Total Filecoin supply.

WINNING_POST_SECTOR_SET_LOOKBACK
Epochs to look back for verifying PoSt proofs.

Traits

NetworkParams
Config trait which handles different network configurations.

Functions

zero_piece_commitment
Returns a zero piece commitment based on the amount of space needed to pad.

Type Definitions

ActorID
Identifier for Actors, includes builtin and initialized actors